Resctructured my  code and added the documentation
