+++ MyTimer/Windows/Headers/finishsignal.h
@@ -7,6 +7,9 @@ namespace Ui {
 class FinishSignal;
 }
 
+/**
+ * @brief The FinishSignal class describes the proper ui window that is dedicated to emit the audio and visual signal after the end of the alarm/timer
+ */
 class FinishSignal : public QDialog
 {
     Q_OBJECT
@@ -16,6 +19,9 @@ public:
     ~FinishSignal();
 
 private slots:
+    /**
+     * @brief on_OK_clicked is implemeted when the button "pushButton" is clicked. It just closes the window.
+     */
     void on_OK_clicked();
 
 private:
