+++ MyTimer/Windows/Sources/finishsignal.cpp
@@ -1,4 +1,4 @@
-#include "finishsignal.h"
+#include "../MyTimer/Windows/Headers/finishsignal.h"
 #include "ui_finishsignal.h"
 
 FinishSignal::FinishSignal(QWidget *parent) :
@@ -13,6 +13,9 @@ FinishSignal::~FinishSignal()
     delete ui;
 }
 
+/**
+ * @brief on_OK_clicked is implemeted when the button "pushButton" is clicked. It just closes the window.
+ */
 void FinishSignal::on_OK_clicked()
 {
     close();
