--- MyTimer/createtimer.h
@@ -1,44 +1,0 @@
-#ifndef CREATETIMER_H
-#define CREATETIMER_H
-
-#include <QDialog>
-#include "chooseaudio.h"
-
-namespace Ui {
-class CreateTimer;
-}
-
-class CreateTimer : public QDialog
-{
-    Q_OBJECT
-
-public:
-    explicit CreateTimer(QWidget *parent = nullptr);
-    ~CreateTimer();
-
-private slots:
-    void on_pushButton_clicked();
-
-    int on_create_clicked();
-
-    void on_chooseAudio_clicked();
-
-    void set_audio_num(int in);
-
-    void on_AmTimer_clicked();
-
-    void on_defaultTimer_clicked();
-
-    void on_alarmTimer_clicked();
-
-signals:
-    void timer_set(int time, int audio_num, bool ifchecked);
-    void timer_alarm_set(int time, int audio_num, bool ifchecked);
-
-private:
-    Ui::CreateTimer *ui;
-    ChooseAudio *CAW;
-    int audio_num;
-};
-
-#endif // CREATETIMER_H
