+++ MyTimer/main.cpp
@@ -1,4 +1,4 @@
-#include "mainwindow.h"
+#include "Windows/Headers/mainwindow.h"
 #include <QApplication>
