--- MyTimer/managetimer.cpp
@@ -1,14 +1,0 @@
-#include "managetimer.h"
-#include "ui_managetimer.h"
-
-ManageTimer::ManageTimer(QWidget *parent) :
-    QDialog(parent),
-    ui(new Ui::ManageTimer)
-{
-    ui->setupUi(this);
-}
-
-ManageTimer::~ManageTimer()
-{
-    delete ui;
-}
